requirements.txt
Django==3.1.12
djangorestframework==3.12.4
djangorestframework-api-key==2.1.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.11.0
django-jazzmin==2.6.0
django-otp==1.1.6
djongo==1.3.7
sqlparse==0.2.4
asgiref==3.3.4
pymongo==3.11.4
mysqlclient==2.0.3
pillow==8.2.0
requests==2.25.1
PyJWT==2.1.0
google-auth==1.30.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
rsa==4.7.2
pytz==2021.1
reportlab==3.6.9
razorpay==1.3.0
pyfcm==1.5.4
weasyprint=67.0